0.3.4
service, dao, DB 수정
- 1) travel 하나 생성하기  → to/: travel이름
- 2) person 하나 생성하기  → to/: PersonBriefly
- 3) event 하나 생성하기  → to/: EventBriefly, event개수, travel기간,
- 4) person 상세 페이지
- 5) event 상세 페이지

+++ back-end/src/main/java/com/spring/mydiv/Dto/Person.java
@@ -89,7 +89,7 @@ public class Person {
 		return Difference;
 	}
 
-	public void setDifference(int difference) {
+	public void setDifference(double difference) {
 		Difference = difference;
 	}
